requirements.txt
opencv-contrib-python
requests
pynmea2
pyserial
torch
imutils